refactor: consolidate interfaces into single interface.ts file
- Consolidated InterfaceEventsAttendedByUserProps and InterfaceUserAddressFieldsProps into interface.ts
- Changed TSDoc to use @property tags for interface properties
- Updated barrel export to reference consolidated interface.ts
- Removed separate type files per CodeRabbit feedback

Follows project pattern used in Post, Event, and User type directories
